+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/ShapeOp.java
@@ -35,7 +35,7 @@ public abstract class ShapeOp extends BaseOp {
 
     public ShapeOp(SameDiff sameDiff,
                    SDVariable i_v,
-                   int[] shape,
+                   long[] shape,
                    boolean inPlace,
                    Object[] extraArgs) {
         super(sameDiff,inPlace,extraArgs);
@@ -68,7 +68,7 @@ public abstract class ShapeOp extends BaseOp {
     }
 
     @Override
-    public List<int[]> calculateOutputShape() {
+    public List<long[]> calculateOutputShape() {
         throw new UnsupportedOperationException();
     }
