+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/executioner/OpExecutioner.java
@@ -351,7 +351,7 @@ public interface OpExecutioner {
      */
     void exec(CustomOp op);
 
-    List<int[]> calculateOutputShape(CustomOp op);
+    List<long[]> calculateOutputShape(CustomOp op);
 
 
     void enableDebugMode(boolean reallyEnable);
